Move misfiled ServerTimeHelper into its own file (Lite)
ServerTimeHelper — an app-wide time-display helper used across ~21 Lite files —
was buried as a top-level class at the top of LocalDataService.Blocking.cs, above
the LocalDataService partial. Moved it verbatim to Lite/Services/ServerTimeHelper.cs.

Same namespace (PerformanceMonitorLite.Services), so zero call-site changes; the
new file carries only the usings it needs (System + PerformanceMonitor.Ui for
TimeDisplayMode). Pure file move, no behavior change.

Solution builds clean; Lite.Tests (544) pass.
